[tvos] Added demo application

+++ platform/darwin/src/MGLLoggingConfiguration.m
@@ -54,7 +54,7 @@
             NSUInteger logTypesCount = sizeof(log_types) / sizeof(os_log_type_t);
             NSAssert(level <= logTypesCount, @"There is an attempt to log a non suported logging level.");
             os_log_type_t logType = log_types[level];
-            os_log_with_type(mapbox_log, logType, "%@ - %d: %@", fileName, line, message);
+            os_log_with_type(mapbox_log, logType, "%@ - %lu: %@", fileName, (unsigned long)line, message);
         } else {
             NSLog(@"[SDK] %@ - %lu: %@", fileName, line, message);
         }
